+++ src/derives/parse.rs
@@ -22,6 +22,7 @@ pub enum ClapAttr {
     Long,
     Flatten,
     Subcommand,
+    Skip,
     Parse(ParserSpec),
     RenameAll(LitStr),
     NameLitStr(Ident, LitStr),
@@ -93,6 +94,7 @@ impl Parse for ClapAttr {
                 "short" => Ok(Short),
                 "flatten" => Ok(Flatten),
                 "subcommand" => Ok(Subcommand),
+                "skip" => Ok(Skip),
                 _ => {
                     let msg = format!("unexpected attribute: {}", name_str);
                     Err(input.error(&msg))