+++ tests/ui/skip_with_other_options.rs
@@ -1,0 +1,15 @@
+use clap::Clap;
+
+#[derive(Clap, Debug)]
+#[clap(name = "test")]
+pub struct Opts {
+    #[clap(long)]
+    a: u32,
+    #[clap(skip, long)]
+    b: u32,
+}
+
+fn main() {
+    let opts = Opts::parse();
+    println!("{:?}", opts);
+}